+++ profiler/lib.profiler.ui/src/org/netbeans/lib/profiler/ui/memory/DiffSampledResultsPanel.java
@@ -190,7 +190,7 @@ public class DiffSampledResultsPanel extends SnapshotSampledResultsPanel {
     }
 
     protected Object computeValueAt(int row, int col) {
-        int index = ((Integer) filteredToFullIndexes.get(row)).intValue();
+        int index = (Integer) filteredToFullIndexes.get(row);
 
         switch (col) {
             case 0:
